+++ src/bacnet/bacdcode.h
@@ -86,6 +86,19 @@ typedef struct BACnetTag {
     uint32_t len_value_type;
 } BACNET_TAG;
 
+typedef struct BACnetCharacterStringBuffer {
+    uint8_t encoding;
+    char *buffer;
+    size_t buffer_size;
+    uint32_t buffer_length;
+} BACNET_CHARACTER_STRING_BUFFER;
+
+typedef struct BACnetOctetStringBuffer {
+    uint8_t *buffer;
+    size_t buffer_size;
+    uint32_t buffer_length;
+} BACNET_OCTET_STRING_BUFFER;
+
 /* max size of a BACnet tag */
 #define BACNET_TAG_SIZE 7
 
@@ -182,6 +195,8 @@ BACNET_STACK_EXPORT
 int encode_application_null(uint8_t *apdu);
 BACNET_STACK_EXPORT
 int encode_context_null(uint8_t *apdu, uint8_t tag_number);
+BACNET_STACK_EXPORT
+int bacnet_null_application_encode(uint8_t *apdu, uint32_t apdu_size);
 BACNET_STACK_EXPORT
 int bacnet_null_application_decode(const uint8_t *apdu, uint32_t apdu_size);
 BACNET_STACK_EXPORT
@@ -385,13 +400,53 @@ bool bacnet_object_id_same(
     uint32_t instance2);
 
 BACNET_STACK_EXPORT
+int encode_application_octet_string_buffer(
+    uint8_t *apdu, const uint8_t *buffer, size_t buffer_size);
+BACNET_STACK_EXPORT
+int encode_context_octet_string_buffer(
+    uint8_t *apdu,
+    uint8_t tag_number,
+    const uint8_t *buffer,
+    size_t buffer_size);
+BACNET_STACK_EXPORT
+int bacnet_octet_string_buffer_application_encode(
+    uint8_t *apdu,
+    uint32_t apdu_size,
+    const uint8_t *buffer,
+    size_t buffer_size);
+BACNET_STACK_EXPORT
+int bacnet_octet_string_buffer_context_encode(
+    uint8_t *apdu,
+    uint32_t apdu_size,
+    uint8_t tag_number,
+    const uint8_t *buffer,
+    size_t buffer_size);
+BACNET_STACK_EXPORT
+int bacnet_octet_string_buffer_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    uint32_t len_value,
+    uint8_t *buffer,
+    size_t buffer_size);
+BACNET_STACK_EXPORT
+int bacnet_octet_string_buffer_application_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    uint8_t *buffer,
+    size_t buffer_size);
+BACNET_STACK_EXPORT
+int bacnet_octet_string_buffer_context_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    uint8_t tag_value,
+    uint8_t *buffer,
+    size_t buffer_size);
+
+BACNET_STACK_EXPORT
 int encode_octet_string(uint8_t *apdu, const BACNET_OCTET_STRING *octet_string);
 BACNET_STACK_EXPORT
 int encode_application_octet_string(
     uint8_t *apdu, const BACNET_OCTET_STRING *octet_string);
-BACNET_STACK_EXPORT
-int encode_application_octet_string_buffer(
-    uint8_t *apdu, const uint8_t *buffer, size_t buffer_size);
 BACNET_STACK_EXPORT
 int encode_context_octet_string(
     uint8_t *apdu, uint8_t tag_number, const BACNET_OCTET_STRING *octet_string);
@@ -430,6 +485,46 @@ uint32_t encode_bacnet_character_string_safe(
     uint8_t encoding,
     const char *pString,
     uint32_t length);
+BACNET_STACK_EXPORT
+uint32_t encode_bacnet_character_string_buffer(
+    uint8_t *apdu, const BACNET_CHARACTER_STRING_BUFFER *value);
+BACNET_STACK_EXPORT
+int encode_context_character_string_buffer(
+    uint8_t *apdu,
+    uint8_t tag_number,
+    const BACNET_CHARACTER_STRING_BUFFER *char_string);
+BACNET_STACK_EXPORT
+int encode_application_character_string_buffer(
+    uint8_t *apdu, const BACNET_CHARACTER_STRING_BUFFER *char_string);
+BACNET_STACK_EXPORT
+int bacnet_character_string_buffer_application_encode(
+    uint8_t *apdu,
+    uint32_t apdu_size,
+    const BACNET_CHARACTER_STRING_BUFFER *value);
+BACNET_STACK_EXPORT
+int bacnet_character_string_buffer_context_encode(
+    uint8_t *apdu,
+    uint32_t apdu_size,
+    uint8_t tag_number,
+    const BACNET_CHARACTER_STRING_BUFFER *value);
+BACNET_STACK_EXPORT
+int bacnet_character_string_buffer_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    uint32_t len_value,
+    BACNET_CHARACTER_STRING_BUFFER *value);
+BACNET_STACK_EXPORT
+int bacnet_character_string_buffer_application_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    BACNET_CHARACTER_STRING_BUFFER *value);
+BACNET_STACK_EXPORT
+int bacnet_character_string_buffer_context_decode(
+    const uint8_t *apdu,
+    uint32_t apdu_size,
+    uint8_t tag_value,
+    BACNET_CHARACTER_STRING_BUFFER *value);
+
 BACNET_STACK_EXPORT
 int encode_bacnet_character_string(
     uint8_t *apdu, const BACNET_CHARACTER_STRING *char_string);